pyproject.toml
[tool.poetry]
name = "taskguardian"
version = "0.1.0"
description = "Taskguardian is a taskwarrior web interface"
authors = ["Bernard Siebens <bernard@siebens.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
tasklib = "^2.5.1"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django-tree-queries = "^0.14.0"
django-ninja = "^0.22.2"
django-ninja-extra = "^0.19.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
